Revised the sections for sessions and running the app.
I more strongly worded the section about the dance keyword in "Running
the application" to further discourage people from using it.

In the sessions section, I listed a couple of session engines that are
suitable for production use to help address any questions about what is
or isn't considered production ready.